Bugfix for OStatus to prevent sending messages from wrong senders
